update reference LHEs
The difference comes from the contact term implementation, which
rearranges some calculation orders.  The ratios old/new in the VBF ME
are something like 1 +/- 1e-18.  The differences in ggH are probably
the same thing, but I had trouble testing since I couldn't reproduce
the bug on MARCC.  If you look at the failed sdiff -s at
https://circleci.com/gh/JHUGen/JHUGen/359, for both ggH and VBF the
xsec and first event are basically identical, because of the tiny difference.
After that it diverges, probably because of vegas.  The first event in ggH
is even more similar than in VBF, and that's probably why on some machines
like MARCC the difference doesn't show up at all.