update reference LHEs
spin 2 changes a tiny bit, similar to what happened for spin 0 in 189755d3.
After 100 events, the momentum is different by a factor of ~(1+1e-6).
We can be pretty confident in attributing this to numerical precision
with the added contact terms.
The other reference files have some changes in the header.